+++ __init__.py
@@ -5,6 +5,7 @@ Created on Wed May 27 10:54:13 2020
 @author: katha
 """
 
-from prisoners_dilemma_exp.prisondersdilemma import run
+from prisoners_dilemma_exp.prisonersdilemma import run
 
-python3 prisoners_dilemma_exp/setup.py install --user
+
+#python3 prisoners_dilemma_exp/setup.py install --user
