--- pleaseendme.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Sun May 31 14:26:09 2020
-
-@author: katha
-"""
-
